moved govActions panel into a separated folder
